etl/dagsterbosch/pyproject.toml
[tool.poetry]
name = "dagsterbosch"
version = "0.0.1"
description = "Dagster server for ETL orchestration"
authors = ["Manuel Caipo <manuel.caipo@boschrexroth.de>"]

[tool.poetry.dependencies]
python = "3.10.11"
dagit = "^1.3.0"
pydantic = {version = "^2.10.6", allow-prereleases = true}
pyyaml = "^6.0.1"
typing-extensions = "^4.12.2"
dagster = "^1.9.10"
dagster-webserver = "^1.9.10"
dagster-postgres = "^0.25.10"
dagster-docker = "^0.25.10"
cx-oracle = "^8.3.0"
paho-mqtt = "^2.1.0"
solace-pubsubplus = "1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
